show seafile logo in the shared link menuitem

+++ fsplugin/build.sh
@@ -3,7 +3,7 @@
 set -x
 set -e
 
-PWD=$(dirname "${BASH_SOURCE[0]}")
+CURRENT_PWD="$(dirname "${BASH_SOURCE[0]}")"
 
 if [ "$(uname -s)" != "Darwin" ]; then
   echo "don't run it if you are not using Mac OS X"
@@ -14,10 +14,14 @@ export CC=$(xcrun -f clang)
 export CXX=$(xcrun -f clang)
 unset CFLAGS CXXFLAGS LDFLAGS
 
-pushd $PWD
-rm -rf CMakeCache.txt CMakeFiles
-cmake -G Xcode -DCMAKE_BUILD_TYPE=Release
+pushd $CURRENT_PWD
+CONFIG=Debug
+if [ a"$1" != "adebug" ]; then
+  rm -rf CMakeCache.txt CMakeFiles
+  CONFIG=Release
+fi
+cmake -G Xcode -DCMAKE_BUILD_TYPE="$CONFIG"
 xcodebuild clean
-xcodebuild -jobs "$(sysctl -n hw.ncpu)" -configuration Release
+xcodebuild -jobs "$(sysctl -n hw.ncpu)" -configuration "$CONFIG"
 popd
